Added kernels for testing the Reduction algorithm.
This is for implementing power-value sensors, and has only been created as a proof of concept. Two kernels have been created:
- Simple Reduction of integers
- Reduction of 'Complex data', laid out in memory as | real | imag | real | imag | ...
The kernels currently use test-data generated within the Class, and have not had their verification methods implemented yet.
There is still a few things that require clarification and/or optimisiation, but this should give the reader an idea of how we intend on this calculation to be implemented.

This corresponds to work on jira-ticket-NGC-47.