Revised code to ensure that the servlet works in manual testing.
For this purposes, I created a `fake_node.py` script in the `debug` folder.

Also made sure that everything passed all of Tyrone's checks, so that I could sleep peacefully at
night knowing that my code complies with the standards.

Still outstanding is the sensor aspect, and unit testing. Still working on that.