Corr3Servlet sensor monitoring unit-test.
Added a unit-test for checking that the sensor values given by the nodes are
seen and reflected by the Corr3Servlet. The test fails (as expected) because
the functionality has not been implemented yet.

Modified the make_a_request() function, because the new test needed it.
Previously this function returned nothing. Now it returns the (reply, informs)
tuple which it gets from actually making a KATCP request. It's now also robust
enough to return None objects if the request fails (which it will do in some
test cases). There may be a better way to do this but this one works for the
time being. Existing tests ignore the output of the function anyway and are not
affected. They still pass.

Also modified the FakeNode class so that it now has a device-status sensor and
a means wherewith the test can modify it to see whether the change propagates
upwards.